个人中心年龄设置;webview联调;离线工单上传接口失败联调;通知滚动添加设置调整调试

+++ OpticalValleyUnite/Classes/UI/VideoPatrol/Model/YQVideoMapPointModel.swift
@@ -29,7 +29,7 @@ class YQVideoMapPointModel: NSObject {
 //    var videoConfigId : Int = 0
     
     //设备摄像头id
-    var equipmentId : Int = 0
+    var equipmentId : Int64 = 0
     
     //新增的是否是巡查过的点
     var isIns : Int = 0
